添加log read&write日志

+++ db/log_format.h
@@ -8,6 +8,8 @@
 #ifndef STORAGE_LEVELDB_DB_LOG_FORMAT_H_
 #define STORAGE_LEVELDB_DB_LOG_FORMAT_H_
 
+#include <cstddef>
+
 namespace leveldb {
 namespace log {
 
@@ -27,10 +29,10 @@ enum RecordType {
 };
 static const int kMaxRecordType = kLastType;
 
-static const int kBlockSize = 32768;
+static const size_t kBlockSize = 32768;
 
 // Header is checksum (4 bytes), length (2 bytes), type (1 byte).
-static const int kHeaderSize = 4 + 2 + 1;
+static const size_t kHeaderSize = 4 + 2 + 1;
 
 }  // namespace log
 }  // namespace leveldb
